Move `Node` and `Tree` into `tree` submodule
